pyproject.toml
[project]
name = "melito"
version = "0.1.0"
description = "Polytonic Ancient Greek OCR pipeline for Melito's Peri Pascha"
requires-python = ">=3.10"
dependencies = [
    "pytesseract>=0.3.10",
    "opencv-python>=4.8",
    "Pillow>=10.0",
    "click>=8.1",
]

[project.scripts]
melito = "melito.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
